More Grappling Hook Debugging
Test (grappling hook) - #3 : Ditched the dot product calculation for using triggers that are above and below the car. Currently trying to find a way to determine where the grapple is during FixedUpdate.

Will need to be returned to later. The two courses of action are get the TurnDirectionGrappling script to function properly or use CheckBox correctly to determine placement.