Implemented Grapple Raycast, Organised More Areas
Feature (grappling hook) - #9 : The grapple raycast now functions as intended. This allows for grappling only on certain layers, and positioning the grapple point on the surface of objects.

Organisation (scripts) - #2 : Some small changes to remove unused code and add comments to parts of functionality. Still more that needs to be done in this area