Tested more solutions for CheckIntersection and Lerp Camera Follow
Test (grappling hook) - #19 : Tested more solutions to the delay issue to no avail. It seems to be an order of execution issue as opposed to a method issue.

Test (camera) - #1 : Created a camera follow lerp script to test the viability of making the camera lag behind the player slightly. Haven't got it fully implemented, but it seems on the right track.