Organised ObjectFollow Scripts, Tested Solution for Finding Direction
Organisation (scripts) - #5 : Compartmentalised the Object Follow Lerp functionality into it's own script. Also removed redundant PivotFunctionDuplicate and made a script reference to the original function in the manager. This allows for the functionality to be more all purpose and used for other objects

Test (grappling hook) - #22 : All the above work was done to apply the lerp functionality to a invisible game object to track direction effectively. This will be trialed after this commit is made