Extra Checks for Follow Object scripts, Tested More Grapple Hook Shooting
Feature (scripts) - #7 : Allowed for extra checks for the object following scripts. Making it so that you can make it either exclusively follow and/or rotate.

Test/Feature (grappling hook) - #29 : Made an object shoot from the car in a direction and attach to walls. Only thing needed to be implemented is making it actually commence the grapple when attached, and sidestep the weird collision issues with raycasts. More information is in a comment in the CheckAndBreakGrapple script